feat: `compilerOptions.types` is set for all kinds
BREAKING CHANGE: in order to disable automatic inclusion of @types packages,
`compilerOptions.types` is set for all project kinds. in the case of
`browser-module`, `browser-executable`, `webworker-module` and
`agnostic-module`, it was previously unset and now it is set to `[]`. In order
to include an @types package, `compilerOptions.types` must override the value
set by the project kind. Extending from any of the `nodejs` kinds, the override
should also include `"node"`. For example, to include the
`@types/webpack-dev-server` package, the value of your
`compilerOptions.types` should be `["node", "webpack-dev-server"]`.